build tests into a separate module
