+++ test/test.js
@@ -1,8 +1,6 @@
 import React                  from "react";
 import JSONInput              from '../src';
 import locale                 from '../src/locale/en'
-import JSONInput_Distribution from '../dist';
-import locale_distribution    from '../dist/locale/en';
 import testSyntaxLogic        from './testSyntaxLogic';
 
 const 
@@ -27,19 +25,6 @@ test('Basic Component Render [SOURCE]', () => {
             test
             id     = 'unique_string'
             locale = {locale}
-        />,
-        { attachTo: window.domNode }
-    );
-    expect(wrapper).toMatchSnapshot();
-});
-
-if(!process.env.DIST_TESTS_DISABLED)
-test('Basic Component Render [DISTRIBUTION]', () => {
-    let wrapper = mount(
-        <JSONInput_Distribution
-            test
-            id     = 'unique_string'
-            locale = {locale_distribution}
         />,
         { attachTo: window.domNode }
     );
@@ -79,42 +64,10 @@ test('All Component Properties Render [SOURCE]', () => {
     expect(wrapper).toMatchSnapshot();
 });
 
-if(!process.env.DIST_TESTS_DISABLED)
-test('All Component Properties Render [DISTRIBUTION]', () => {
-    let wrapper = mount(
-        <JSONInput_Distribution
-            test
-            id                = 'unique_string'
-            locale            = {locale_distribution}
-            placeholder       = {{
-                hello : 'world',
-                foo   : 'bar'
-            }}
-            viewOnly          = {false}
-            onChange          = {sampleFunctions.getResults}
-            confirmGood       = {true}
-            height            = '500px'
-            width             = '100%'
-            onKeyPressUpdate  = {false}
-            waitAfterKeyPress = {600}
-            modifyErrorText   = {sampleFunctions.modWarning}
-            theme             = 'light_mitsuketa_tribute'
-            colors            = {{
-                string : '#DAA520'
-            }}
-            style             = {{
-                labelColumn : {
-                    display : 'hidden'
-                },
-                contentBox : {
-                    pointerEvents : 'none'
-                }
-            }}
-        />,
-        { attachTo: window.domNode }
-    );
-    expect(wrapper).toMatchSnapshot();
-});
+if(!process.env.DIST_TESTS_DISABLED){
+    var testBuild = require('./testBuild.js');
+    testBuild();
+}
 
 sampleData = {
     basic : {
